refactor: move createCars to InputView
