+++ src/main/java/ui/InputView.java
@@ -1,23 +1,30 @@
 package ui;
 
-import java.util.Scanner;
+import domain.Car;
+
+import java.util.*;
 
 public class InputView {
+    private static final String NAME_DELIMITER = ",";
 
-    private Scanner scanner;
-
-    public InputView() {
-        this.scanner = new Scanner(System.in);
-    }
-
-    public String getName() {
-        System.out.println("경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분).");
-        return scanner.nextLine();
-    }
-
-    public int getTryCount() {
+    public static int getTryCount() {
         System.out.println("시도할 횟수는 몇 회 인가요?");
+        Scanner scanner = new Scanner(System.in);
         return scanner.nextInt();
+    }
+
+    public static List<Car> getCars() {
+        System.out.println("경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분).");
+        Scanner scanner = new Scanner(System.in);
+        return createCars(scanner.nextLine());
+    }
+
+    static List<Car> createCars(String names) {
+        List<Car> cars = new ArrayList<>();
+        for (String name : Arrays.asList(names.split(NAME_DELIMITER))) {
+            cars.add(new Car(name));
+        }
+        return Collections.unmodifiableList(cars);
     }
 
 }
