+++ src/test/java/domain/CarsTest.java
@@ -8,19 +8,9 @@ import java.util.Arrays;
 import static org.assertj.core.api.Assertions.assertThat;
 
 public class CarsTest {
-
-    @Test
-    public void createCarsTest() {
-        String names = "a,b,c";
-        Cars cars = new Cars(names);
-        assertThat(cars.getCars()).hasSize(3);
-        assertThat(cars.getCars()).contains(new Car("a"), new Car("b"), new Car("c"));
-    }
-
     @Test
     public void moveAllTest() {
-        String names = "a,b,c";
-        Cars cars = new Cars(names);
+        Cars cars = new Cars(Arrays.asList(new Car("a"), new Car("b"), new Car("c")));
         MovableStrategy trueMovableStrategy = () -> true;
 
         cars.moveAll(trueMovableStrategy);
